Move to TypeScript build
