+++ test/composition-helper-test.js
@@ -1,5 +1,5 @@
 var assert = require('chai').assert;
-var Terminal = require('../src/xterm');
+var Terminal = require('../build/xterm');
 
 describe('CompositionHelper', function () {
   var terminal;
