+++ test/escape-sequences-test.js
@@ -2,7 +2,7 @@ var glob = require('glob');
 var fs = require('fs');
 var pty = require('pty.js');
 var sleep = require('sleep');
-var Terminal = require('../src/xterm');
+var Terminal = require('../build/xterm');
 
 var CONSOLE_LOG = console.log;
