+++ typings.json
@@ -1,0 +1,6 @@
+{
+  "name": "xterm",
+  "globalDependencies": {
+    "node": "registry:env/node#6.0.0+20160918225031"
+  }
+}
